Python/Lab/_gDrive/_GYetki/requirements.txt
httplib2
oauth2client
PyDrive
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv